+++ zarrs_data_type/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: `DataType::metadata_fill_value()` is now fallible
 - **Breaking**: `DataType::{identifier,size,fixed_size}()` are no longer `const`
 - **Breaking**: `DataType::from_metadata()` now returns a `PluginCreateError` on error instead of `UnsupportedDataTypeError`
+- **Breaking**: `DataType::from_metadata()` has an additional `ExtensionAliasesDataTypeV3` parameter
 - **Breaking**: `DataType::[fixed_]size()` are no longer `const`
 - Bump `derive_more` to 0.2.0
 - Bump `half` to 2.3.1
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - **Breaking**: Remove `UnsupportedDataTypeError`
 - **Breaking**: Remove `DataType::identifier()`
+- **Breaking**: Remove `TryFrom<DataTypeMetadataV3>` for `DataType`
 
 ## [0.1.0] - 2025-01-24